+++ ggml/src/ggml-metal/ggml-metal-device.cpp
@@ -518,7 +518,7 @@ ggml_metal_pipeline_with_params ggml_metal_library_get_pipeline_ssm_conv_batched
 }
 
 ggml_metal_pipeline_with_params ggml_metal_library_get_pipeline_ssm_scan(ggml_metal_library_t lib, const ggml_tensor * op)  {
-    GGML_TENSOR_LOCALS( int32_t, ne0, op->src[0], ne);
+    GGML_TENSOR_LOCALS( int64_t, ne0, op->src[0], ne);
 
     char base[256];
     char name[256];
@@ -586,11 +586,11 @@ ggml_metal_pipeline_with_params ggml_metal_library_get_pipeline_gated_delta_net(
     char name[256];
 
     // v is src[2], dimensions: S_v = ne[0], H = ne[1]
-    const int ne20 = (int)op->src[2]->ne[0]; // S_v
-    const int ne21 = (int)op->src[2]->ne[1]; // H
-    const int ne30 = (int)op->src[3]->ne[0]; // G
+    const int ne20 = op->src[2]->ne[0]; // S_v
+    const int ne21 = op->src[2]->ne[1]; // H
+    const int ne30 = op->src[3]->ne[0]; // G
 
-    const int nsg = (int)(op->src[2]->ne[0]/32);
+    const int nsg = op->src[2]->ne[0]/32;
 
     GGML_ASSERT(op->src[5]->type == GGML_TYPE_F32);
     GGML_ASSERT(op->ne[0] == ne20 * ne21);
@@ -621,8 +621,8 @@ ggml_metal_pipeline_with_params ggml_metal_library_get_pipeline_solve_tri(ggml_m
     char name[256];
 
     const int nsg = 8;
-    const int n   = (int)op->src[1]->ne[1];
-    const int k   = (int)op->src[1]->ne[0];
+    const int n   = op->src[1]->ne[1];
+    const int k   = op->src[1]->ne[0];
 
     snprintf(base, 256, "kernel_solve_tri_%s", ggml_type_name(op->src[0]->type));
     snprintf(name, 256, "%s_nsg=%d_n=%d_k=%d", base, nsg, n, k);
@@ -700,8 +700,8 @@ ggml_metal_pipeline_with_params ggml_metal_library_get_pipeline_mul_mm(ggml_meta
 }
 
 ggml_metal_pipeline_with_params ggml_metal_library_get_pipeline_mul_mv(ggml_metal_library_t lib, const ggml_tensor * op) {
-    GGML_TENSOR_LOCALS( int32_t, ne0, op->src[0], ne);
-    GGML_TENSOR_LOCALS( int32_t, ne1, op->src[1], ne);
+    GGML_TENSOR_LOCALS( int64_t, ne0, op->src[0], ne);
+    GGML_TENSOR_LOCALS( int64_t, ne1, op->src[1], ne);
 
     char base[256];
     char name[256];
@@ -929,8 +929,8 @@ ggml_metal_pipeline_with_params ggml_metal_library_get_pipeline_mul_mm_id(ggml_m
 }
 
 ggml_metal_pipeline_with_params ggml_metal_library_get_pipeline_mul_mv_id(ggml_metal_library_t lib, const ggml_tensor * op) {
-    GGML_TENSOR_LOCALS( int32_t, ne0, op->src[0], ne);
-    GGML_TENSOR_LOCALS( int32_t, ne1, op->src[1], ne);
+    GGML_TENSOR_LOCALS( int64_t, ne0, op->src[0], ne);
+    GGML_TENSOR_LOCALS( int64_t, ne1, op->src[1], ne);
 
     char base[256];
     char name[256];
@@ -1335,8 +1335,8 @@ ggml_metal_pipeline_with_params ggml_metal_library_get_pipeline_flash_attn_ext(
     const int32_t dk = (int32_t) op->src[1]->ne[0];
     const int32_t dv = (int32_t) op->src[2]->ne[0];
 
-    const int32_t ns10 = (int32_t)(op->src[1]->nb[1]/op->src[1]->nb[0]);
-    const int32_t ns20 = (int32_t)(op->src[2]->nb[1]/op->src[2]->nb[0]);
+    const int64_t ns10 = op->src[1]->nb[1]/op->src[1]->nb[0];
+    const int64_t ns20 = op->src[2]->nb[1]/op->src[2]->nb[0];
 
     // do bounds checks for the mask?
     const bool bc_mask = op->src[3] && (op->src[3]->ne[1] % 8 != 0);
@@ -1401,8 +1401,8 @@ ggml_metal_pipeline_with_params ggml_metal_library_get_pipeline_flash_attn_ext_v
     const int32_t dk = (int32_t) op->src[1]->ne[0];
     const int32_t dv = (int32_t) op->src[2]->ne[0];
 
-    const int32_t ns10 = (int32_t)(op->src[1]->nb[1]/op->src[1]->nb[0]);
-    const int32_t ns20 = (int32_t)(op->src[2]->nb[1]/op->src[2]->nb[0]);
+    const int64_t ns10 = op->src[1]->nb[1]/op->src[1]->nb[0];
+    const int64_t ns20 = op->src[2]->nb[1]/op->src[2]->nb[0];
 
     snprintf(base, 256, "kernel_%s_%s_dk%d_dv%d",
             "flash_attn_ext_vec",
